# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloudLens API",
    "description": "Azure FinOps Managed Service \u2014 cost intelligence API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Audit",
        "description": "Query the audit trail (most recent first).",
        "operationId": "get_audit_api_v1_admin_audit_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "_system",
              "title": "Tenant Id"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditRecord"
                  },
                  "title": "Response Get Audit Api V1 Admin Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/compliance/audit-integrity/{tenant_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Audit Integrity",
        "description": "Verify the audit hash chain is intact \u2014 proof the trail is tamper-evident.",
        "operationId": "audit_integrity_api_v1_admin_compliance_audit_integrity__tenant_id__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Audit Integrity Api V1 Admin Compliance Audit Integrity  Tenant Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/compliance/matrix": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Compliance Matrix",
        "description": "The SOC 2-aligned control matrix with per-control CLI evidence commands.",
        "operationId": "compliance_matrix_api_v1_admin_compliance_matrix_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Compliance Matrix Api V1 Admin Compliance Matrix Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/compliance/evidence-export": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Evidence Export",
        "description": "Generate a compliance evidence pack: the control matrix with CLI proof\ncommands, plus the audit-chain integrity result. Records the export itself\nin the audit log (CC7.2). The admin page renders/downloads this.",
        "operationId": "evidence_export_api_v1_admin_compliance_evidence_export_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "_system",
              "title": "Tenant Id"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "admin",
              "title": "Actor"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evidence Export Api V1 Admin Compliance Evidence Export Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditAction": {
        "type": "string",
        "enum": [
          "login",
          "access_denied",
          "tenant_scope_violation",
          "rate_limited",
          "tenant_created",
          "tenant_updated",
          "tenant_deleted",
          "budget_created",
          "budget_updated",
          "budget_deleted",
          "alert_rule_created",
          "alert_rule_updated",
          "alert_rule_deleted",
          "waste_resolved",
          "ingest_triggered",
          "report_generated",
          "report_downloaded",
          "evidence_exported"
        ],
        "title": "AuditAction"
      },
      "AuditRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "audit_record"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "action": {
            "$ref": "#/components/schemas/AuditAction"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "actor_type": {
            "type": "string",
            "title": "Actor Type",
            "default": "user"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id",
            "default": ""
          },
          "outcome": {
            "type": "string",
            "title": "Outcome",
            "default": "success"
          },
          "source_ip": {
            "type": "string",
            "title": "Source Ip",
            "default": ""
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "prev_hash": {
            "type": "string",
            "title": "Prev Hash",
            "default": ""
          },
          "record_hash": {
            "type": "string",
            "title": "Record Hash",
            "default": ""
          },
          "ttl": {
            "type": "integer",
            "title": "Ttl",
            "default": 63072000
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "action",
          "actor"
        ],
        "title": "AuditRecord"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}